+++ NEWS.md
@@ -3,6 +3,10 @@ Phan NEWS
 ??? ?? 2020, Phan 3.2.3 (dev)
 -----------------------
 
+Bug fixes:
++ Properly analyze the right hand side for `$cond || throw ...;` (e.g. emit `PhanCompatibleThrowException`) (#4199)
++ Don't infer implications of `left || right` on the right hand expression when the right hand side has no side effects. (#4199)
+
 Sep 19 2020, Phan 3.2.2
 -----------------------